requirements.txt
wmctrl-python3
pyhk
